CI test

+++ scripts/bundle_umbrella.sh
@@ -4,6 +4,7 @@
 repo_path=$1
 repo_name=$(basename $repo_path)
 repo_url=$(git config --get remote.origin.url)
+bundle_path=$repo_path/$repo_name.bundle
 
 # vendor the umbrella repository
 echo "Bundling the $repo_name repository"
@@ -14,11 +15,16 @@ echo "Commit Ref Name: $CI_COMMIT_REF_NAME"
 
 echo "Cleaning existing path $repo_path"
 rm -rf $repo_path
+
 echo "Cloning $repo_url to $repo_path"
 git clone $repo_url $repo_path
-cd $repo_path
+
 echo "Bundling repository at HEAD"
-git bundle create ../$repo_name.bundle HEAD
-cd -
+cd $repo_path
+git bundle create $bundle_path HEAD
+
+test -f $bundle_path || { echo "Error: $bundle_path does not exist"; exit 1; }
+
 echo "Deleting repository $repo_path"
+cd -
 rm -rf $repo_path
